Fix #55, remove the <div> around images
Images used to have a single parent, and the attribute fixup used that fact for saving attributes. After this change many images are part of the same parent so that doesn't work anymore. Instead we just expect the images to stay in the same order after being uploaded to Gmail.