Fix #172, ask template question and start loading overlay before collecting information from tabs
Note you can still end up on the 'Preparing your email' screen for a while,
but the loading screen appears in the meantime

Also fix #114, delay creating iframe until document.body is for sure present.

This adds a new pattern for creating a promise with exposed resolve/reject,
which makes it easier to await on global information.